--- src/lib/components/query-param-select.svelte
@@ -1,45 +1,0 @@
-<script lang="ts">
-  import { goto } from '$app/navigation';
-  import { page } from '$app/stores';
-  import { updateQueryParameters } from '$lib/utilities/update-query-parameters';
-
-  export let label: string;
-  export let options: string[] | Record<string, string>;
-  export let value: string;
-  export let parameter: string = null;
-
-  let _value = (parameter && $page.query.get(parameter)) || value;
-
-  const id = `${parameter || label}-filter`;
-
-  $: {
-    updateQueryParameters({
-      parameter,
-      value: _value,
-      query: $page.query,
-      path: $page.path,
-      goto,
-    });
-  }
-</script>
-
-<div class="flex flex-col items-start justify-center">
-  <label for={id} class="text-gray-600 text-xs whitespace-nowrap">
-    {label}
-  </label>
-  <select
-    class="block border-2 text-base p-2 w-full h-10"
-    {id}
-    bind:value={_value}
-  >
-    {#if Array.isArray(options)}
-      {#each options as value}
-        <option {value}>{value}</option>
-      {/each}
-    {:else}
-      {#each Object.entries(options) as [label, value]}
-        <option {value}>{label}</option>
-      {/each}
-    {/if}
-  </select>
-</div>
